refactor: 폴더 구분
