+++ src/main/java/com/sejong/archiveservice/core/common/file/Filepath.java
@@ -1,4 +1,4 @@
-package com.sejong.archiveservice.core.common;
+package com.sejong.archiveservice.core.common.file;
 
 
 import java.util.Objects;
