+++ src/main/java/com/sejong/archiveservice/infrastructure/archive/converter/FilepathsConverter.java
@@ -1,6 +1,6 @@
 package com.sejong.archiveservice.infrastructure.archive.converter;
 
-import com.sejong.archiveservice.core.common.Filepaths;
+import com.sejong.archiveservice.core.common.file.Filepaths;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
